Extract functions that help checking current mode
