+++ packages/wp-now/src/wp-playground-wordpress/index.ts
@@ -1,1 +1,4 @@
 export * from './is-valid-wordpress-version';
+export * from './is-plugin-directory';
+export * from './is-theme-directory';
+export * from './is-wp-content-directory';
